Add files via upload
Sorry, this one is a bit of a nightmare: it will only work on my computer for directories of images. I'm looking to fix that to make it use local resources for image loading. The big fix here was for the character class. The remaining issues are loading the correct images into the character at the correct time.